Use PGXS for installation and un-installation
Notably, this allows it to work whether or not there is a `doc`
directory and allows the `DESTDIR` variable to work, which is often used
when building packages.

Work around an empty `doc` directory by only setting `DOCS` if there are
docs. Add the `uninstall_more` target and add it to the dependencies for
PGXS's `install` target to clean up bin and doc files.